backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiologger>=0.7.0",
    "asyncio>=3.4.3",
    "fastapi[standard]>=0.115.6",
    "httpx>=0.28.1",
    "motor>=3.7.0",
    "protobuf>=5.29.2",
    "python-dotenv>=1.0.1",
    "redis[hiredis]>=5.2.1",
    "setuptools>=75.6.0",
    "upstox-python-sdk>=2.9.0",
    "websockets>=14.1",
]
